Refactor to comply with codeclimate
